Start migration vuex store
Vuex manages a lot of the communication with girder, so there's a lot to
sift through here.

This is enough to load the experiments list, click on a scan, and load
the Dataset view. I actively avoided any refactoring in favor of getting
things working temporarily.

Vuex refers to scans as "sessions" and images as "datasets".

I also disabled the login monitor for now.