Remove unnecessary parent IDs
The un-flattened API required you to have the ID of every parent of an
object you wanted to use the API on: if you wanted to GET an image, you
needed the ID of the image, the scan, the experiment, and the session.

The newly flattened API only requires you to have the ID of the object
in question, so we can avoid passing around all those parent IDs.